+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQM5RB3DREMXCME0R9EXJBBC"
+        "01KQRDRW6MG290H0AZ4JFAZEKN"
     )
     r"""The unique identifier of the data source"""