+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQM5RB3FJK3FP878PTNNTMCE"
+        "01KQRDRW6XH4GZV4VHG9A501N2"
     )
     r"""The unique identifier of the data source"""