+++ packages/orq-rc/RELEASES.md
@@ -7509,3 +7509,13 @@ Based on:
 - [python v4.9.0-rc.18] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.18] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.18 - packages/orq-rc
+
+## 2026-05-04 02:43:41
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.11 (2.881.17) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.19] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.19] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.19 - packages/orq-rc